+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## [7.4.0]
+* `registerSingleton` now returns the passed instance as a return value thanks to the PR by @Rexios80 https://github.com/fluttercommunity/get_it/pull/242
+* In some cases GetIt will print error messages to the debug output. Now this won't happen anymore in release mode and can be completely disabled by setting `GetIt.noDebugOutput=true` following the PR from @James1345 
 ## [7.3.0]
 
 New features: